refactor: 回答ドメインの定義を概念別に分割する
Answer に関する型が 1 ファイルに集まって責務を追いにくくなっていたため、公開 API を維持したまま author、content、entry、settings、label などの単位へ整理する。
