+++ server/domain/src/form/answer/title.rs
@@ -1,0 +1,13 @@
+use deriving_via::DerivingVia;
+use types::non_empty_string::NonEmptyString;
+
+#[derive(Clone, DerivingVia, Default, Debug, PartialEq)]
+#[deriving(From, Into, IntoInner, Serialize(via: Option::<NonEmptyString>), Deserialize(via: Option::<NonEmptyString>
+))]
+pub struct AnswerTitle(Option<NonEmptyString>);
+
+impl AnswerTitle {
+    pub fn new(title: Option<NonEmptyString>) -> Self {
+        Self(title)
+    }
+}
